blas: Use indices for children
Instead of using pack and handles, use a vector to store the nodes of
the Blas. First node in the vector is root, second node is unused. A node
has just the index of its left child, then its right child index is just
left_child_index + 1. Notice that since index 1 is unused, we use this
value as a marker for no children.