blas: Use ranges
Simplify BVH nodes by using ranges (offset, count) instead of owning
the vector of primitives. These primitives are stored in the BLAS object.

Moreover, if the node has primitives it means it is a leaf. Otherwise,
the left child index is store in triangles.offset and the right child
index is just the left node index plus one.